refactor: Restructure the project to align with Flutter recommended project structure
