+++ lib/data/services/rest_client.dart
@@ -1,6 +1,6 @@
 import 'package:dio/dio.dart';
+import 'package:lessons_tasks_assignment/data/repositories/lessons/models/raw_lesson.dart';
 import 'package:lessons_tasks_assignment/domain/task.dart';
-import 'package:lessons_tasks_assignment/repositories/lessons/models/raw_lesson.dart';
 import 'package:retrofit/retrofit.dart';
 
 part 'rest_client.g.dart';
