+++ lib/features/lessons/lessons_route.dart
@@ -2,12 +2,12 @@ import 'package:dio/dio.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
+import 'package:lessons_tasks_assignment/data/repositories/lessons/interceptors/json_interceptor.dart';
+import 'package:lessons_tasks_assignment/data/repositories/lessons/lessons_repository.dart';
+import 'package:lessons_tasks_assignment/data/services/rest_client.dart';
 import 'package:lessons_tasks_assignment/features/lesson/lesson_route.dart';
 import 'package:lessons_tasks_assignment/features/lessons/cubit/lessons_cubit.dart';
 import 'package:lessons_tasks_assignment/features/lessons/view/lessons_view.dart';
-import 'package:lessons_tasks_assignment/repositories/lessons/interceptors/json_interceptor.dart';
-import 'package:lessons_tasks_assignment/repositories/lessons/lessons_repository.dart';
-import 'package:lessons_tasks_assignment/repositories/lessons/rest_client.dart';
 
 part 'lessons_route.g.dart';
