+++ test/repositories/lessons/lessons_repository_test.dart
@@ -1,11 +1,11 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:lessons_tasks_assignment/data/repositories/lessons/lessons_repository.dart';
+import 'package:lessons_tasks_assignment/data/repositories/lessons/models/raw_lesson.dart';
 import 'package:lessons_tasks_assignment/domain/answer.dart';
 import 'package:lessons_tasks_assignment/domain/content_component.dart';
 import 'package:lessons_tasks_assignment/domain/lesson.dart';
 import 'package:lessons_tasks_assignment/domain/page.dart';
 import 'package:lessons_tasks_assignment/domain/task.dart';
-import 'package:lessons_tasks_assignment/repositories/lessons/lessons_repository.dart';
-import 'package:lessons_tasks_assignment/repositories/lessons/models/raw_lesson.dart';
 import 'package:mockito/mockito.dart';
 
 import '../../mocks.mocks.dart';
